+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## [3.1.0] - 2020-07-19
 
 ### Added